Activity-Lang: wire Job invocation in the activity::Term builder

+++ tests/vault/gear/scheduler-activity-test.cpp
@@ -146,7 +146,7 @@ namespace test {
         {
           ActivityDetector detector;
           
-          size_t x1=rand(), x2=rand();
+          uint64_t x1=rand(), x2=rand();
           Time nomTime = lib::test::randTime();
           Activity feed{x1,x2};
           Activity feed2{x1+1,x1+2};
